path updates - P1
