+++ src/App.tsx
@@ -1,14 +1,14 @@
-import ListNav from "components/ListNav/ListNav";
-import MyLists from "components/MyLists/MyLists";
-import { useState } from "react";
-import { createBrowserRouter, RouterProvider } from "react-router-dom";
-import "./App.css";
 import UserAuthFactory from "auth/UserAuthFactory";
 import AllIngredients from "components/AllIngredients/AllIngredients";
 import Home from "components/Home/Home";
 import Layout from "components/Layout/Layout";
-import LoggedIn from "components/LoggedIn/LoggedIn";
+import ListNav from "components/ListNav/ListNav";
+import MyLists from "components/MyLists/MyLists";
+import Welcome from "components/Welcome/Welcome";
+import { useState } from "react";
+import { createBrowserRouter, RouterProvider } from "react-router-dom";
 import BackendFactory from "services/BackendFactory";
+import "./App.css";
 
 function App() {
 	const [isFirstLoggin, setIsFirstLoggin] = useState(true);
@@ -18,16 +18,16 @@ function App() {
 	const router = createBrowserRouter([
 		{
 			path: "/",
-			element: <Home userAuth={userAuth} />,
+			element: <Welcome userAuth={userAuth} />,
 			errorElement: <div>404 Page Not Found</div>,
 		},
 		{
 			element: <Layout userAuth={userAuth} />,
 			children: [
 				{
-					path: "/logged-in",
+					path: "/home",
 					element: (
-						<LoggedIn
+						<Home
 							userAuth={userAuth}
 							isFirstLoggin={isFirstLoggin}
 							setIsFirstLoggin={(val) => setIsFirstLoggin(val)}
