--- src/components/Home/Home.css
@@ -1,27 +1,0 @@
-@import "assets/Common.css";
-
-.container {
-	display: flex;
-	flex-direction: column;
-	height: 100vh;
-}
-
-.container > .sub-color {
-	background-color: var(--sub-color); 
-	width: 100%;
-	flex-grow: 1; 
-}
-
-.header {
-	position: sticky;
-	top: 0;
-	z-index: 1;
-	text-align: center;
-	width: 100%;
-	padding: 10px 0;
-	color: white;
-	font-size: 1.5rem;
-	font-weight: bold;
-	text-transform: uppercase;
-	box-shadow: 0px 4px 6px rgba(0, 0, 0, 0.1);
-}
