+++ src/components/Home/Home.tsx
@@ -1,68 +1,51 @@
+/**
+ * This page is what the user is redirected to once theyve successfully logged in.
+ */
+import { useEffect, useState } from "react";
 import { useNavigate } from "react-router-dom";
 import { UserAuth } from "../../auth/UserAuth";
+import Loading from "../Loading/Loading";
 import "./Home.css";
 
 interface HomeProps {
-	userAuth: UserAuth; // Receive UserAuth from props
+	userAuth: UserAuth;
+	isFirstLoggin: boolean;
+	setIsFirstLoggin: (val: boolean) => void;
 }
 
-function Home({ userAuth }: HomeProps) {
+function Home({ userAuth, isFirstLoggin, setIsFirstLoggin }: HomeProps) {
 	const navigate = useNavigate();
+	const [isLoading, setIsLoading] = useState(true);
 
-	const handleLogin = () => {
-		if (!userAuth.isAuthenticated()) {
-			userAuth.login(); // Log in the user using Auth0 or Mock User
-		}
-		if (userAuth.isAuthenticated()) {
-			navigate("/logged-in");
-		}
-	};
+	useEffect(() => {
+		const checkAuth = async () => {
+			while (userAuth.isProcessing()) {
+				await new Promise((resolve) => setTimeout(resolve, 100)); // Wait 100ms before rechecking
+			}
 
+			if (userAuth.isAuthenticated() && isFirstLoggin) {
+				userAuth.completeLogin();
+				setIsFirstLoggin(false);
+			} else if (!userAuth.isAuthenticated()) {
+				navigate("/");
+			}
+			setIsLoading(false);
+		};
 
+		setIsLoading(true);
+		checkAuth();
+	}, [navigate, userAuth, isFirstLoggin, setIsFirstLoggin]);
 
 	return (
-		<div className="container">
+		<div
+			className="sub-color"
+			style={{ height: "100%" }}
+		>
 			{/* Background Logo */}
 			<div className="logo-background" />
 
-			{/* Banner */}
-			<span className="header-color header">Welcome to Cupboard!</span>
-
 			{/* Main Content */}
-			<div className="sub-color"></div>
-			<div
-				style={{
-					position: "absolute", // Change to absolute positioning
-					bottom: "20px", // Set bottom offset
-					left: "50%", // Center horizontally
-					transform: "translateX(-50%)", // Center alignment
-					zIndex: 1,
-				}}
-			>
-				{/* Button */}
-				<button
-					onClick={handleLogin}
-					className="primary-color"
-					style={{
-						position:"absolute",
-						top: "-175px",
-						left: "50%",
-						transform: "translateX(-50%)",
-						padding: "15px 30px",
-						fontSize: "1.2rem", // Use rem for responsive font size
-						color: "white",
-						border: "none",
-						borderRadius: "5px",
-						cursor: "pointer",
-						textTransform: "uppercase",
-						letterSpacing: "1px",
-						minWidth: "200px", // Ensure a minimum width for the button
-						maxWidth: "90%", // Ensure the button doesn’t overflow on smaller screens
-					}}
-				>
-					Log In
-				</button>
-			</div>
+			{isLoading && true ? <Loading /> : <div></div>}
 		</div>
 	);
 }
